+++ es/components/CheckboxField.js
@@ -1,4 +1,4 @@
 import Checkbox from "antd/lib/checkbox";
-import { createComponent } from "./BaseComponent";
+import { createComponent } from "./createComponent";
 import eventMap from "../maps/eventMap";
 export default createComponent(Checkbox, eventMap);
