+++ es/components/NumberField.js
@@ -1,4 +1,4 @@
 import InputNumber from "antd/lib/input-number";
 import mapError from "../maps/mapError";
-import createComponent from "./BaseComponent";
+import createComponent from "./createComponent";
 export default createComponent(InputNumber, mapError);
