+++ es/components/RadioField.js
@@ -1,4 +1,4 @@
 import eventMap from "../maps/eventMap";
 import { RadioField as Radio } from "./MultiSelect";
-import createComponent from "./BaseComponent";
+import createComponent from "./createComponent";
 export default createComponent(Radio, eventMap);
