+++ es/components/SelectField.js
@@ -5,7 +5,7 @@ function _objectSpread(target) { for (var i = 1; i < arguments.length; i++) { va
 function _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }
 
 import { customMap } from "../maps/mapError";
-import createComponent from "./BaseComponent";
+import createComponent from "./createComponent";
 import { SelectField as Select } from "./MultiSelect";
 var selectFieldMap = customMap(function (mapProps, _ref) {
   var _ref$input = _ref.input,
