+++ lib/components/CheckboxField.js
@@ -7,12 +7,12 @@ exports["default"] = void 0;
 
 var _checkbox = _interopRequireDefault(require("antd/lib/checkbox"));
 
-var _BaseComponent = require("./BaseComponent");
+var _createComponent = require("./createComponent");
 
 var _eventMap = _interopRequireDefault(require("../maps/eventMap"));
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
 
-var _default = (0, _BaseComponent.createComponent)(_checkbox["default"], _eventMap["default"]);
+var _default = (0, _createComponent.createComponent)(_checkbox["default"], _eventMap["default"]);
 
 exports["default"] = _default;
