+++ lib/components/CheckboxGroupField.js
@@ -9,7 +9,7 @@ var _checkbox = _interopRequireDefault(require("antd/lib/checkbox"));
 
 var _mapError = require("../maps/mapError");
 
-var _BaseComponent = require("./BaseComponent");
+var _createComponent = require("./createComponent");
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
 
@@ -32,6 +32,6 @@ var checkboxGroupMap = (0, _mapError.customMap)(function (mapProps, _ref) {
   });
 });
 
-var _default = (0, _BaseComponent.createComponent)(CheckboxGroup, checkboxGroupMap);
+var _default = (0, _createComponent.createComponent)(CheckboxGroup, checkboxGroupMap);
 
 exports["default"] = _default;
