+++ lib/components/DatePicker.js
@@ -11,7 +11,7 @@ var _datePicker = _interopRequireDefault(require("antd/lib/date-picker"));
 
 var _mapError = require("../maps/mapError");
 
-var _BaseComponent = require("./BaseComponent");
+var _createComponent = require("./createComponent");
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
 
@@ -59,9 +59,9 @@ var datePickerMapRU = (0, _mapError.customMap)(function (mapProps, _ref2) {
     format: displayFormat
   });
 });
-var DatePickerFieldRU = (0, _BaseComponent.createComponent)(_datePicker["default"], datePickerMapRU);
+var DatePickerFieldRU = (0, _createComponent.createComponent)(_datePicker["default"], datePickerMapRU);
 exports.DatePickerFieldRU = DatePickerFieldRU;
-var DatePickerField = (0, _BaseComponent.createComponent)(_datePicker["default"], datePickerMap);
+var DatePickerField = (0, _createComponent.createComponent)(_datePicker["default"], datePickerMap);
 exports.DatePickerField = DatePickerField;
-var MonthPickerField = (0, _BaseComponent.createComponent)(MonthPicker, datePickerMap);
+var MonthPickerField = (0, _createComponent.createComponent)(MonthPicker, datePickerMap);
 exports.MonthPickerField = MonthPickerField;
