+++ lib/components/LazyTextField.js
@@ -7,7 +7,7 @@ exports["default"] = void 0;
 
 var _input = _interopRequireDefault(require("antd/lib/input"));
 
-var _BaseComponent = _interopRequireDefault(require("./BaseComponent"));
+var _createComponent = _interopRequireDefault(require("./createComponent"));
 
 var _mapError = require("../maps/mapError");
 
@@ -48,6 +48,6 @@ var bluredFieldMap = function bluredFieldMap(_ref) {
   });
 };
 
-var _default = (0, _BaseComponent["default"])(_input["default"], bluredFieldMap);
+var _default = (0, _createComponent["default"])(_input["default"], bluredFieldMap);
 
 exports["default"] = _default;
