+++ lib/components/RadioField.js
@@ -9,10 +9,10 @@ var _eventMap = _interopRequireDefault(require("../maps/eventMap"));
 
 var _MultiSelect = require("./MultiSelect");
 
-var _BaseComponent = _interopRequireDefault(require("./BaseComponent"));
+var _createComponent = _interopRequireDefault(require("./createComponent"));
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
 
-var _default = (0, _BaseComponent["default"])(_MultiSelect.RadioField, _eventMap["default"]);
+var _default = (0, _createComponent["default"])(_MultiSelect.RadioField, _eventMap["default"]);
 
 exports["default"] = _default;
