+++ lib/components/SelectField.js
@@ -7,7 +7,7 @@ exports["default"] = void 0;
 
 var _mapError = require("../maps/mapError");
 
-var _BaseComponent = _interopRequireDefault(require("./BaseComponent"));
+var _createComponent = _interopRequireDefault(require("./createComponent"));
 
 var _MultiSelect = require("./MultiSelect");
 
@@ -44,6 +44,6 @@ var selectFieldMap = (0, _mapError.customMap)(function (mapProps, _ref) {
   });
 });
 
-var _default = (0, _BaseComponent["default"])(_MultiSelect.SelectField, selectFieldMap);
+var _default = (0, _createComponent["default"])(_MultiSelect.SelectField, selectFieldMap);
 
 exports["default"] = _default;
