+++ lib/components/SliderField.js
@@ -9,7 +9,7 @@ var _slider = _interopRequireDefault(require("antd/lib/slider"));
 
 var _mapError = require("../maps/mapError");
 
-var _BaseComponent = _interopRequireDefault(require("./BaseComponent"));
+var _createComponent = _interopRequireDefault(require("./createComponent"));
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
 
@@ -37,6 +37,6 @@ var sliderMap = (0, _mapError.customMap)(function (mapProps, _ref) {
   });
 });
 
-var _default = (0, _BaseComponent["default"])(_slider["default"], sliderMap);
+var _default = (0, _createComponent["default"])(_slider["default"], sliderMap);
 
 exports["default"] = _default;
