+++ lib/components/SwitchField.js
@@ -9,7 +9,7 @@ var _switch = _interopRequireDefault(require("antd/lib/switch"));
 
 var _mapError = require("../maps/mapError");
 
-var _BaseComponent = _interopRequireDefault(require("./BaseComponent"));
+var _createComponent = _interopRequireDefault(require("./createComponent"));
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
 
@@ -38,6 +38,6 @@ var switchMap = (0, _mapError.customMap)(function (mapProps, _ref) {
   });
 });
 
-var _default = (0, _BaseComponent["default"])(_switch["default"], switchMap);
+var _default = (0, _createComponent["default"])(_switch["default"], switchMap);
 
 exports["default"] = _default;
