+++ lib/components/TextField.js
@@ -9,10 +9,10 @@ var _input = _interopRequireDefault(require("antd/lib/input"));
 
 var _textFieldMap = _interopRequireDefault(require("../maps/textFieldMap"));
 
-var _BaseComponent = _interopRequireDefault(require("./BaseComponent"));
+var _createComponent = _interopRequireDefault(require("./createComponent"));
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
 
-var _default = (0, _BaseComponent["default"])(_input["default"], _textFieldMap["default"]);
+var _default = (0, _createComponent["default"])(_input["default"], _textFieldMap["default"]);
 
 exports["default"] = _default;
