+++ lib/index.js
@@ -6,7 +6,7 @@ Object.defineProperty(exports, "__esModule", {
 Object.defineProperty(exports, "createComponent", {
   enumerable: true,
   get: function get() {
-    return _BaseComponent.createComponent;
+    return _createComponent.createComponent;
   }
 });
 Object.defineProperty(exports, "customMap", {
@@ -94,7 +94,7 @@ Object.defineProperty(exports, "TextField", {
   }
 });
 
-var _BaseComponent = require("./components/BaseComponent");
+var _createComponent = require("./components/createComponent");
 
 var _mapError = require("./maps/mapError");
