+++ src/components/CheckboxField.js
@@ -1,5 +1,5 @@
 import Checkbox from "antd/lib/checkbox";
-import { createComponent } from "./BaseComponent";
+import { createComponent } from "./createComponent";
 import eventMap from "../maps/eventMap";
 
 export default createComponent(Checkbox, eventMap);
