+++ src/components/TextField.js
@@ -1,5 +1,5 @@
 import Input from "antd/lib/input";
 import textFieldMap from "../maps/textFieldMap";
-import createComponent from "./BaseComponent";
+import createComponent from "./createComponent";
 
 export default createComponent(Input, textFieldMap);
